@@ -10,7 +10,7 @@
   "scripts": {
     "lint": "jshint . && eslint .",
     "validate": "npm ls",
-    "start": "node node_modules/react-native/local-cli/cli.js start"
+    "start": "react-native start"
   },
   "dependencies": {
     "events": "^1.1.1",
@@ -18,11 +18,10 @@
     "hark": "1.1.3",
     "inherits": "^2.0.1",
     "merge": "^1.2.0",
-    "react": "16.3.1",
-    "react-dom": "15.4.2",
-    "react-native": "0.55.4",
-    "react-native-permissions": "~1.1.1",
-    "react-native-webrtc": "1.63.0",
+    "react": "17.0.1",
+    "react-native": "0.64.0",
+    "react-native-permissions": "^3.0.5",
+    "react-native-webrtc": "1.89.3",
     "sdp-translator": "^0.1.15",
     "ua-parser-js": "^0.7.7",
     "uuid": "~2.0.1"
